refactor: created forgot-password component
